Fixed Search and temporary fix for Add Customer
I simply added an else statement for adding customer so if the inputs
are blanks or just spaces then the page simply reload.
Some alternatives I thought of and still figuring out:
1. after submitting blank inputs, the page reload with a message onto
the body of the page that the Customer was not added because some/all
inputs are blanks; the form they submitted is returned back, but they
will have to click the add button to view the modal again.

2. after submission, implement ajax and jquery to inform user if the
inputs are invalid (blank/spaces) with custom error messages appearing
on the modal next to the corresponding field input.

3. use js/jquery to indicate errors while user are still typing into the
field and submission button won't work until errors are clear.