Changed Search form method  to GET and changed Add to an ajax request
- deleted customer_form.html and form_template because not using them.
- switched search form method from POST to GET so that user can use the
back button to go to previous results
- made a separate function for add in views.py
- made the add form an ajax request;
- and error messages will appear in the modal in the <div id =
"msg"></div>  if submitted fields are blanks; I wanted errors to appear
inline with the field, but I'm still figuring to do that with forms.Form
so this will do if I cannot do it.

Other stuff that I'm going to add:
- 404 pages (not sure what this was exactly about, so I'll restudy it)
- restrict names to only letters (a-z,Z-A)
- fix form validations
- make page better looking if I can/update its look/update texts (such
as error messages)

Other stuff that I'm thinking of adding if I'm able to:
- add an edit button for editing Customer
- change search and remove
- change search url, after user submit input, to a more simple one such
as /CustomerData/search=first instead of the long one with
csrfmiddlewaretoken in it.
- make <div></div> holding the table list to change its content to
customer details if customer is clicked; the url will change from
/CustomerData/searchsajflsjsjfsomething to
/CustomerData/searchsajflsjsjfsomething/43 with 43 as the pk or id of
the Customer
- if the header "First Name" is clicked
the list will be sort by first name from a->z
if clicked again z->a
but to go back to the original unsorted list, must click List
button...so I'll think more about this.
- implement the same above for last name
- add time_created attribute for Customer and then add time_created to
the table list. So if the time_created header is clicked, the table will
sort the Customer from most recent to least, etc.
- ^ I don't think I'll need time_created, since the Customer's pk
already tells which one was created first, but I'll put time_created if
I want to be specific about the date they joined/were created.